Refactor & Cleanup
This PR will allow us to do further work on the init option with a
cleaner CLI. Renamed a test, we should have more of those eventually.
Removed a dependency we are not going to use and added a more modular
approach to building up the options object passed to the compiler when
the CLI command isn’t init.