Resolve Errors in Parser Validating the wrong option
Parser: Improvements to core to allow flexibility

Fix issues with parser and work in progress to make the parser reusable
for multiple services.

Parser: Allow more flexibility and combine pkg resolving

Parser: Allow flexibility in build

Fix: Bump dependencies for testability

feat: Add comments for easier fixtures.

Added comments that will allow a PR to be easier in the future.