Parser: Integrate Parser (#59)
Integrates the structure of the parser being reusable. 
Inquirer is seperated into two instances, one if the user 
only runs with the init flag and another if the user runs with packages as arguments. 
Packages are validated and checked for existency. 
The answers of an inquirer prompt is checked if empty or not and later sent down 
to the parser module where we will later create a webpack configuration file out of transformation rules.
The parser module is reusable, meaning it can be rerun multiple times after an package has been validated.
This may be slow when running with many packages, so we should keep an eye on that.