feat: initial work on new init structure
Changes the structure to the init feature a bit. We’re now making the
parser validate everything, as we can later run the inquirer prompt
after running yeoman, reading from a json file. This will allow us to
have a more fluid process and removes a lot of «trash» code. Disabled
listing for now, as this is work in progress.