Feature/yeoman generators (#93)
* fix: remove unneeded code

* feat: initial work on new init structure

Changes the structure to the init feature a bit. We’re now making the
parser validate everything, as we can later run the inquirer prompt
after running yeoman, reading from a json file. This will allow us to
have a more fluid process and removes a lot of «trash» code. Disabled
listing for now, as this is work in progress.

* feat: Generate yeoman instance from a package

Makes it possible to make a inquirer prompt with a remote package.

* feat: try using webpack-chain for obj-creation

* feat: get build ready for refactor

* fix: rename parser to creator

* fix: enable listing for lib library and fix issues

* fix: don't disable lint for destructuring

* fix: remove inquirer prompt and fix minor issues

Removes inquirer, we’re not using it explicitly anymore. Fixed some
review mistakes.

* fix: remove destruct and local import in some tests