fix: default named import bug
There are issues when we import TS compiled JS files into JS because TS compiles default exports to `default` named import/export which makes it not possible to import inside JS directly.
Some of the ways to fix this is using .default in all calls or avoid using default export/import totally.

Ref: https://basarat.gitbooks.io/typescript/docs/tips/defaultIsBad.html