chore: Remove redundant multiple property
The multiple property of the optionDefinitions object is set to true when the option expects multiple values to be passed via the command line. By default this property is false. There is no need to specify this explicitly. All the following optionDefinitions which do not expect multiple values don't specify it only this one does.