fix: updated getPackageManager logic
The getPackageManager function of the package-manager module, which was supposed to resolve the
default Package Manager. The expected behaviour is that if yarn is present, yarn is used else npm is
used. However, the logic didnt allow the ideal behaviour. As per the updated logic, we check for
local yarn. If present, we use yarn, else npm is used. Also, the spawn statement is eliminated and
so are the related test statements.

ISSUES CLOSED: #941