fix: cleanup `package-utils` package (#1822)
This commit removes the redundant `package-utils` package and moves the shared outing utils to the core CLI. This is because we felt the package was redundant and this seemed like the better way of doing things. All the tests and code has been migrated to core CLI and any packages which use it directly read it from the main CLI package.